Router hardening + tx credits refactor

+++ include/mmx/Router.h
@@ -65,7 +65,7 @@ private:
 		bool is_outbound = false;
 		uint32_t height = 0;
 		uint32_t credits = 0;
-		uint32_t tx_credits = 0;
+		uint64_t tx_credits = 0;
 		int32_t ping_ms = 0;
 		int64_t last_receive_ms = 0;
 		int64_t connected_since_ms = 0;
